+++ pandora/__init__.py
@@ -112,7 +112,7 @@ def run(
     # and the machine will store the necessary parameters for the following scale
     for scale in range(pandora_machine.num_scales):  # pylint:disable=unused-variable
         # Trigger the machine step by step
-        for elem in list(cfg)[1:]:
+        for elem in list(cfg["pipeline"])[1:]:
             pandora_machine.run(elem, cfg)
             # If the machine gets to the begin state, pass to next scale
             if pandora_machine.state == "begin":
@@ -203,7 +203,7 @@ def main(cfg_path: str, output: str, verbose: bool) -> None:
         img_right,
         disp_min,
         disp_max,
-        cfg["pipeline"],
+        cfg,
         disp_min_right,  # type: ignore
         disp_max_right,  # type: ignore
     )